changes to export tools

+++ export_tools.py
@@ -5,27 +5,24 @@ import rdkit.Chem as chem
 import rdkit.Chem.AllChem as ac
 from functools import partial
 from multiprocessing import Pool
+import itertools as it
 
 from .classes import StructUnit, FGInfo
 from .convenience_tools import MolFileError
 from .optimization import *
 
 
-def fg_prune(input_folder, output_folder, fg, fg_num):
+def fg_prune(ifolder, fg, fg_num):
     """
-    Copies molecules with a given functional group between folders.
+    Deletes molecules without a given functional group from a folder.
     
     Parameters
     ----------
-    input_folder : str
-        The full path of the folder holding V3000 .mol files. Any
-        molecules with a functional group `fg` in this folder are copied
-        to `output_folder`.
-        
-    output_folder : str
-        The full path of the folder into which files with a given 
-        functional group `fg` are copied to.
-    
+    ifolder : str
+        The full path of the folder holding .mol files. Any
+        molecules without functional group `fg` `fg_num` amount of times
+        are deleted.
+
     fg : str
         The name of the functional group which the copied molecules must
         possess. The name must correspond to one of the name of a 
@@ -42,23 +39,20 @@ def fg_prune(input_folder, output_folder, fg, fg_num):
         
     """
     
-    for file_name in os.listdir(input_folder):
-        path = os.path.join(input_folder, file_name)
-        try:
-            mol = StructUnit(path, minimal=True)
-        
-        except MolFileError as error:
-            print('V3000 {}.'.format(path))
+    for file_name in os.listdir(ifolder):
+        if not file_name.endswith('.mol'):
             continue
+        path = os.path.join(ifolder, file_name)
+        mol = StructUnit(path, minimal=True)
     
         mol.func_grp = next((x for x in 
                                 FGInfo.functional_group_list if 
                                 x.name == fg), None)
         mol.heavy_ids = []
         mol._generate_heavy_attrs()
-        if len(mol.find_functional_group_atoms()) == fg_num:
-            print('Copying {}.'.format(path))
-            shutil.copy(path, output_folder)
+        if len(mol.find_functional_group_atoms()) != fg_num:
+            print('Deleting {}.'.format(path))
+            os.remove(path)
 
 def fg_distance_prune(folder, fg):
     """
@@ -77,24 +71,23 @@ def fg_distance_prune(folder, fg):
     
     for file_name in os.listdir(folder):
         path = os.path.join(folder, file_name)
-        try:
-            mol = StructUnit(path, minimal=True)
-            
-        except MolFileError as error:
-            print('V3000 {}.'.format(path))
+        if not file_name.endswith('.mol'):
             continue
-    
+
+        mol = StructUnit(path, minimal=True)
+
         mol.func_grp = next((x for x in 
                                 FGInfo.functional_group_list if 
                                 x.name == fg), None)
         
         mol.heavy_ids = []
         mol._generate_heavy_attrs()
-        
         g = mol.graph('heavy')
-        if nx.shortest_path_length(g, *mol.heavy_ids) < 3:
-            print('Removing {}.'.format(path))
-            os.remove(path)
+        for start, end in it.combinations(mol.heavy_ids, 2):
+            if nx.shortest_path_length(g, *mol.heavy_ids) < 3:
+                print('Removing {}.'.format(path))
+                os.remove(path)
+                break
 
 def substurct_prune(folder, substruct):
     """
@@ -107,15 +100,17 @@ def substurct_prune(folder, substruct):
         substructure and deleted.
         
     substruct : str
-        The smiles string of the substructure, which if present in a 
+        The smarts string of the substructure, which if present in a 
         molecule causes it to be deleted from `folder`.
     
     """
     
-    substruct_mol = chem.MolFromSmiles(substruct)
+    substruct_mol = chem.MolFromSmarts(substruct)
     for file_name in os.listdir(folder):
+        if not file_name.endswith('.mol'):
+            continue
         path = os.path.join(folder, file_name)
-        mol = chem.MolFromMolFile(path)
+        mol = chem.MolFromMolFile(path, sanitize=False, removeHs=False)
         if mol.HasSubstructMatch(substruct_mol):
             print('Removing {}.'.format(path))
             os.remove(path)
@@ -247,7 +242,7 @@ def optimize_folder(path, macromodel_path):
         macro_mol.prist_mol_file = name
 
     md_opt = partial(macromodel_md_opt, macromodel_path=macromodel_path, 
-        timeout=False, temp=1000, confs=1000, eq_time=100, sim_time=10000)    
+        timeout=False, temp=300, confs=500, eq_time=50, sim_time=500)    
     
     with Pool() as p:
         p.map(md_opt, macro_mols)
